Update project file to VS Community 2017 or later and fix type error

+++ libvorbis.cpp
@@ -569,7 +569,7 @@ const double rate_mapping_44_stereo[12]=
 
 float vorbis_quality_to_bitrate(const float& q)
 {
-	float ds = 0.0, _is;
+	double ds = 0.0, _is;
 	int is = 0.0;
 
 	ds =modf(q, &_is);
